cors/redirect-preflight.htm

<!DOCTYPE html>
<meta charset=utf-8>
<title>CORS - redirect with preflight</title>
<meta name=author title="Odin Hørthe Omdal" href="mailto:odiho@opera.com">
<meta name=author title="Fernando Jiménez Moreno" href="mailto:ferjmoreno@gmail.com">

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js?pipe=sub></script>

<h1>Redirect with preflight</h1>

<div id=log></div>
<script>

// Request count for cache busting and easy identifying of request in traffic
// analyzer.
var req_c = 0;

var CROSSDOMAIN_URL = CROSSDOMAIN + 'resources/cors-makeheader.py?';

/*
 * Redirection with preflights.
 */
function redir_preflight(code) {
    test(function() {
        var client = new XMLHttpRequest();
        var redirect =
          encodeURIComponent(CROSSDOMAIN_URL + 'headers=x-test&' + req_c++);

        client.open('GET',
                    CROSSDOMAIN_URL + 'headers=x-test&location=' + redirect
                                    + '&code=' + code + '&preflight=' + code
                                    + '&' + req_c++,
                    false);
        client.setRequestHeader('x-test', 'test');
        assert_throws(null, function() { client.send(null) });

    }, 'Redirect ' + code + ' on preflight');
}
redir_preflight(301);
redir_preflight(302);
redir_preflight(303);
redir_preflight(307);
redir_preflight(308);

/*
 * Redirection after successfull (200) preflight.
 */
function redir_after_preflight(code) {
    test(function() {
        var client = new XMLHttpRequest();
        var redirect = encodeURIComponent(
            CROSSDOMAIN + 'resources/cors-makeheader.py?headers=x-test&' + req_c++
        );

        client.open('GET', CROSSDOMAIN + 'resources/cors-makeheader.py?'
                           + 'preflight=200&headers=x-test&location='
                           + redirect + '&code=' + code + '&' + req_c++,
                    false);
        client.setRequestHeader('x-test', 'test');
        client.send(null);
        assert_equals(client.status, 200, "Successfull redirect");

    },
    'Allow redirect ' + code + ' after succesful (200) preflight');
}
redir_after_preflight(301);
redir_after_preflight(302);
redir_after_preflight(303);
redir_after_preflight(307);
redir_after_preflight(308);

</script>
